test: improves testing strategy for config values.
