+++ src/hypertrace/agent/__init__.py
@@ -78,7 +78,7 @@ class Agent:
         if not self.is_enabled():
             return
         try:
-            self._init.flask_init(app)
+            self._init.init_instrumentation_flask(app)
             self._init.dump_config()
         except Exception as err:  # pylint: disable=W0703
             logger.error(constants.EXCEPTION_MESSAGE,
