--- src/hypertrace/agent/config/agent-config.yaml
@@ -1,29 +1,0 @@
-service_name: pythonagent_001
-reporting:
-  endpoint: http://localhost:9411/api/v2/spans1
-  secure: true
-  trace_reporter_type: ZIPKIN
-  token: "TestToken"
-  opa:
-    endpoint: "https://opa.traceableai:8181/"
-    poll_period_seconds: 50
-    enabled: true
-data_capture:
-  http_headers:
-    request: false
-    response: false
-  http_body:
-    request: false
-    response: false
-  rpc_metadata:
-    request: false
-    response: false
-  rpc_body:
-    request: false
-    response: false
-  body_max_size_bytes: 123457
-propagation_formats: ["TRACECONTEXT","B3"]
-_use_console_span_exporter: true
-enabled: false
-resource_attributes: {}
-
